Split out code related to example subcommand to separate module
